requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
httpie==3.2.2
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
openapi-codec==1.3.2
packaging==23.2
psycopg2-binary==2.9.9
Pygments==2.17.2
PySocks==1.7.1
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
